Move stdlib into separate module
