+++ tests/fixtures/pass/stdlib/stdlib_http_request_response_layer_vm.modes
@@ -1,0 +1,1 @@
+vm
