+++ tests/fixtures/pass/stdlib/stdlib_http_request_response_layer_vm.stdout
@@ -1,0 +1,15 @@
+3
+q
+hello world
+two
+/home
+true
+hello world
+two
+/home
+host
+x-test
+second
+201
+ok
+HTTP/1.1 201 Created
