+++ tests/fixtures/pass/stdlib/stdlib_net_socket_controls_vm.modes
@@ -1,0 +1,1 @@
+vm
