+++ tests/fixtures/pass/stdlib/stdlib_net_socket_controls_vm.stdout
@@ -1,0 +1,5 @@
+string
+string
+true
+x
+true
